13. January 2022, Santa Detection/requirements.txt
albumentations==0.5.2
torch>=1.7.1
numpy>=1.19.5
pandas>=1.2.4
torchvision>=0.8.2
PyYAML==5.4.1
tqdm==4.56.0
opencv-python
scikit-learn
scikit-image
python-json-logger>=0.1.11
#efficientnet-pytorch
